requirements.txt
# Main dependencies
livekit-agents>=1.2.18
livekit-plugins-openai>=1.2.0
livekit-plugins-cartesia>=0.4.0
livekit-plugins-noise-cancellation>=0.2.0
livekit-plugins-silero>=1.2.0
livekit-plugins-turn-detector>=1.2.0
python-dotenv

# Development dependencies
pytest>=8.3.4
pytest-asyncio>=0.24.0
ruff>=0.8.4
